Repeat request in case of connection reset.
Connection reset issue is pretty common but can be bypassed if we
keep repeating the request. Initially, when any kind of Exception was thrown
we used to give up. But now we are making an exception for the connection
reset error. Twenty attempts will be made to overcome this error for each new
request.