Repeat request upto twenty times in case of connection reset.
Connection reset error is pretty common after the internet providers
started banning sites like thekickasstorrents. But due to some mis-
configurations, it is possible to bypass the connection reset error
by simply repeating the request. This commit is doing just that.
We confirm that the IOException is for connection being reset and then
repeat the same request upto twenty times.

The implementation is untested because for some odd reason every request
to kickasstorrents.to seems to be failing with a request timeout error.
It does not look like this commit has screwed things up because the problem
persists in earlier commits, even in origin/master. On the other hand, the
mcat binary in my pc is failing with the right error (connection reset) so
I am really skeptic about the current state of mediacat. This needs debugging
and once we start to get connection reset while making requests, we must test
this commit.